requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.4
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
graphql-core==3.2.7
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
lia-web==0.2.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
motor==3.7.1
packaging==25.0
passlib==1.7.4
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
strawberry-graphql==0.287.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
